sale order detail API

+++ core-entity/src/main/java/vn/viettel/core/db/entity/promotion/PromotionSaleProduct.java
@@ -20,7 +20,7 @@ public class PromotionSaleProduct extends BaseEntity {
     @Column(name ="PRODUCT_ID")
     private Long productId;
     @Column(name ="QUANTITY")
-    private Float quantity;
+    private Integer quantity;
     @Column(name ="STATUS")
     private Integer status;
 }
